@@ -16,14 +16,14 @@
 		"bootstrap-icons": "^1.11.3"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.2.5",
-		"@sveltejs/kit": "^2.7.0",
-		"@sveltejs/vite-plugin-svelte": "^3.1.2",
-		"svelte": "^4.2.19",
+		"@sveltejs/adapter-auto": "^3.3.1",
+		"@sveltejs/kit": "^2.7.3",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"svelte": "^5.1.3",
 		"svelte-check": "^4.0.5",
-		"tslib": "^2.7.0",
+		"tslib": "^2.8.0",
 		"typescript": "^5.6.3",
-		"vite": "^5.4.8"
+		"vite": "^5.4.10"
 	},
 	"type": "module"
 }
